cleanup directory structure
