--- tools/hub/.gitignore
@@ -1,2 +1,0 @@
-hub
-*.o
