--- tools/hub/Makefile
@@ -1,19 +1,0 @@
-TARGET = hub
-
-CFLAGS = -Wall -O2
-LDFLAGS= -pthread
-CC=gcc
-
-
-SRCS := hub.c
-OBJS = $(subst .c,.o,$(SRCS))
-
-${TARGET}: ${OBJS}
-	$(CC) ${LDFLAGS} -o $@ $^
-
-.c.o:
-	$(CC) ${CFLAGS} -o $@ -c $<
-
-.PHONY: clean
-clean:
-	rm -f ${TARGET} ${OBJS}
