--- tools/mgdump/.gitignore
@@ -1,1 +1,0 @@
-mgdump
