+++ tools/pipe.c
@@ -17,8 +17,7 @@
 #include <pthread.h>
 #include <signal.h>
 
-
-#include "../../kmod/hpio.h"
+#include <hpio.h>
 
 #define MAX_CPU	64
 #define BULKNUM	128
@@ -26,7 +25,7 @@
 struct pipe_thread {
 	pthread_t	tid;
 
-	char *devpath_r, *devpath_l;
+	char *devpath_r, *devpath_w;
 	
 	int rd_fd;	/* fd for read from hpio device */
 	int wr_fd;	/* fd for write to output file */
@@ -56,8 +55,8 @@ int count_online_cpus(void)
 void usage(char *progname)
 {
 	printf("%s usage:\n"
-	       "\t -r: right hpio device path\n"
-	       "\t -l: left hpio device path\n",
+	       "\t -r: read hpio device path\n"
+	       "\t -w: write hpio device path\n",
 		progname);
 }
 
@@ -82,8 +81,7 @@ void *pipe_body(void *arg)
 	CPU_ZERO(&target_cpu_set);
 	CPU_SET(pdt->cpu, &target_cpu_set);
 	pthread_setaffinity_np(pdt->tid, sizeof(cpu_set_t), &target_cpu_set);
-	printf("thread %d on cpu %d: %s->%s\n", pdt->cpu, sched_getcpu(),
-	       pdt->devpath_r, pdt->devpath_l);
+	printf("thread %d on cpu %d\n", pdt->cpu, sched_getcpu());
 
 	/* initialize packet buffer */
 	for (n = 0; n < BULKNUM; n++) {
@@ -108,7 +106,7 @@ void *pipe_body(void *arg)
 			perror("writev");
 
 		printf("read %d pkt from %s, write %d pkt to %s on cpu %d\n",
-		       cnt, pdt->devpath_r, ret, pdt->devpath_l, pdt->cpu);
+		       cnt, pdt->devpath_r, ret, pdt->devpath_w, pdt->cpu);
 	}
 
 	printf("thread %d finished\n", pdt->cpu);
@@ -121,20 +119,19 @@ void *pipe_body(void *arg)
 int main (int argc, char **argv)
 {
 	int rd_fd, wr_fd, n, ncpus, ret, ch;
-	char *devpath_r, *devpath_l;
+	char *devpath_r, *devpath_w;
 	struct pipe_thread pdts[MAX_CPU];
-	struct pipe_thread rev_pdts[MAX_CPU];
 
 	devpath_r = NULL;
-	devpath_l = NULL;
+	devpath_w = NULL;
 
-	while ((ch = getopt(argc, argv, "r:l:")) != -1) {
+	while ((ch = getopt(argc, argv, "r:w:")) != -1) {
 		switch(ch) {
 		case 'r' :
 			devpath_r = optarg;
 			break;
-		case 'l' :
-			devpath_l = optarg;
+		case 'w' :
+			devpath_w = optarg;
 			break;
 		default:
 			usage(argv[0]);
@@ -142,7 +139,7 @@ int main (int argc, char **argv)
 		}
 	}
 
-	if (!devpath_r || !devpath_l) {
+	if (!devpath_r || !devpath_w) {
 		printf("two hpio devices must be specified\n");
 		usage(argv[0]);
 		return -1;
@@ -153,7 +150,6 @@ int main (int argc, char **argv)
 
 	
 	/* open read hpio descriptor and dispatch to pipe_thread */
-	/* devpath_r -> devpath_l */
 	for (n = 0; n < ncpus; n++) {
 
 		rd_fd = open(devpath_r, O_RDONLY);
@@ -163,9 +159,9 @@ int main (int argc, char **argv)
 			return -1;
 		}
 
-		wr_fd = open(devpath_l, O_WRONLY);
+		wr_fd = open(devpath_w, O_WRONLY);
 		if (wr_fd < 0) {
-			fprintf(stderr, "cannot open device %s\n", devpath_l);
+			fprintf(stderr, "cannot open device %s\n", devpath_w);
 			perror("open");
 			return -1;
 		}
@@ -173,7 +169,7 @@ int main (int argc, char **argv)
 		pdts[n].rd_fd = rd_fd;
 		pdts[n].wr_fd = wr_fd;
 		pdts[n].devpath_r = devpath_r;
-		pdts[n].devpath_l = devpath_l;
+		pdts[n].devpath_w = devpath_w;
 		pdts[n].cpu = n;
 
 		ret = pthread_create(&pdts[n].tid, NULL, pipe_body, &pdts[n]);
@@ -182,39 +178,6 @@ int main (int argc, char **argv)
 			return -1;
 		}
 	}
-
-	/* create reverse path threads */
-	/* devpath_l -> devpath_r */
-	for (n = 0; n < ncpus; n++) {
-
-		rd_fd = open(devpath_l, O_RDONLY);
-		if (rd_fd < 0) {
-			fprintf(stderr, "cannot open device %s\n", devpath_r);
-			perror("open");
-			return -1;
-		}
-
-		wr_fd = open(devpath_r, O_WRONLY);
-		if (wr_fd < 0) {
-			fprintf(stderr, "cannot open device %s\n", devpath_l);
-			perror("open");
-			return -1;
-		}
-
-		rev_pdts[n].rd_fd = rd_fd;
-		rev_pdts[n].wr_fd = wr_fd;
-		rev_pdts[n].devpath_r = devpath_l;
-		rev_pdts[n].devpath_l = devpath_r;
-		rev_pdts[n].cpu = n;
-
-		ret = pthread_create(&rev_pdts[n].tid, NULL, pipe_body,
-				     &rev_pdts[n]);
-		if (ret < 0) {
-			perror("pthread_create");
-			return -1;
-		}
-	}
-
 
 
 	/* set signal */
@@ -227,7 +190,6 @@ int main (int argc, char **argv)
 	// thread join
 	for (n = 0; n < ncpus; n++) {
 		pthread_join(pdts[n].tid, NULL);
-		pthread_join(rev_pdts[n].tid, NULL);
 	}
 
 	return 0;
