--- tools/pipe/.gitignore
@@ -1,2 +1,0 @@
-pipe
-*.o
