--- tools/pipe/Makefile
@@ -1,19 +1,0 @@
-TARGET = pipe
-
-CFLAGS = -Wall -O2
-LDFLAGS= -pthread
-CC=gcc
-
-
-SRCS := pipe.c
-OBJS = $(subst .c,.o,$(SRCS))
-
-${TARGET}: ${OBJS}
-	$(CC) ${LDFLAGS} -o $@ $^
-
-.c.o:
-	$(CC) ${CFLAGS} -o $@ -c $<
-
-.PHONY: clean
-clean:
-	rm -f ${TARGET} ${OBJS}
